Tests/LibWeb/Text/expected/canvas/2d-draw-invalidates-cached-command.txt

cached command has new generation: true
